Fix redirects for organizers & move other organizer views to correct file
