+++ studygroups/views/organizer.py
@@ -5,6 +5,7 @@ from django.template import RequestContext
 from django.template.loader import render_to_string
 from django.core.urlresolvers import reverse, reverse_lazy
 from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
 from django.core.exceptions import PermissionDenied
 from django.core.mail import EmailMultiAlternatives, send_mail
 from django.contrib import messages
@@ -26,6 +27,7 @@ from studygroups.models import report_data
 from studygroups.models import generate_all_meetings
 from studygroups.models import get_team_users
 from studygroups.forms import StudyGroupForm
+from studygroups.forms import FacilitatorForm
 from studygroups.decorators import user_is_organizer
 
 
@@ -53,6 +55,35 @@ def organize(request):
         'today': timezone.now(),
     }
     return render_to_response('studygroups/organize.html', context, context_instance=RequestContext(request))
+
+
+class FacilitatorCreate(CreateView):
+    model = User
+    form_class = FacilitatorForm
+    success_url = reverse_lazy('studygroups_organize')
+
+    def form_valid(self, form):
+        self.object = form.save(commit=False)
+        self.object.email = self.object.username
+        self.object.save()
+        facilitator = Facilitator(user=self.object)
+        facilitator.save()
+        # TODO - send password reset email to facilitator
+        return http.HttpResponseRedirect(self.get_success_url())
+
+
+class FacilitatorUpdate(UpdateView):
+    model = User
+    form_class = FacilitatorForm
+    success_url = reverse_lazy('studygroups_organize')
+    context_object_name = 'facilitator' # Need this to prevent the SingleObjectMixin from overriding the user context variable used by the auth system
+
+
+class FacilitatorDelete(DeleteView):
+    model = User
+    success_url = reverse_lazy('studygroups_organize')
+    template_name = 'studygroups/confirm_delete.html'
+    context_object_name = 'facilitator' # Need this to prevent the SingleObjectMixin from overriding the user context variable used by the auth system
 
 
 class CourseUpdate(UpdateView):
